upload FireBase

+++ src/app/screen/admin/student/student.component.ts
@@ -1,6 +1,8 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { StudentService } from 'src/app/services/student.service';
+import Swal from 'sweetalert2';
+
 @Component({
   selector: 'app-student',
   templateUrl: './student.component.html',
@@ -20,10 +22,28 @@ export class StudentComponent implements OnInit {
       })
   }
   remove(item: any) {
-    this.studentService.removeStudent(item.id)
-      .subscribe(data => {
-        this.students = this.students.filter(students => students.id != item.id);
-      })
+    Swal.fire({
+      title: `Bạn có muốn xóa tài khoản ${item.name}`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Xóa',
+      cancelButtonText: 'Hủy' 
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Xóa thành công!',
+          'Your file has been deleted.',
+          'success'
+        )
+        this.studentService.removeStudent(item.id)
+        .subscribe(data => {
+          this.students = this.students.filter(students => students.id != item.id);
+        })
+      }
+    })
+ 
   }
   search(){
     this.getStudent(this.keyword);
